Testing the code against old BSM3G analyzer.  Will write README soon

+++ src/DataBinner.h
@@ -13,7 +13,7 @@ using namespace std;
 
 class DataPiece {
  public:
-  vector<int> data;
+  vector<float> data;
   double begin, end, width;
   int bins, Nfold;
   string name;
